Java API to create Wizard Dialogs
A wizard dialog is constructed with a number of panels, and each panel
contains user-configurable components such as radio buttons, sliders, or
text fields. The idea here is that the application user, by pressing
either the Next or Back buttons, can "flip" across these panels,
entering information on each one until he or she completes the path. At
this point, the Next button often changes to a Finish button, and when
pressed, the dialog will close. The instantiating class should then be
able to retrieve all of the data that was entered (saved in the context
object). At any time, the user can also press the Cancel button to quit
the dialog and discard any data that has been entered so far.